feat(s2n-quic): add console perf event subscriber
