+++ quic/s2n-quic/Cargo.toml
@@ -24,6 +24,9 @@ provider-address-token-default = [
     "zerocopy-derive",
     "zeroize",
 ]
+provider-event-console-perf = [
+    "humansize"
+]
 provider-event-tracing = ["s2n-quic-core/event-tracing"]
 provider-tls-default = ["s2n-quic-tls-default"]
 provider-tls-rustls = ["s2n-quic-rustls"]
@@ -55,6 +58,7 @@ cfg-if = "1"
 cuckoofilter = { version = "0.5", optional = true }
 futures = { version = "0.3", default-features = false, features = ["std"] }
 hash_hasher = { version = "2", optional = true }
+humansize = { version = "2", optional = true }
 rand = "0.8"
 rand_chacha = "0.3"
 s2n-codec = { version = "=0.6.0", path = "../../common/s2n-codec" }
